Move network request plugin to new package
