+++ packages/platforms/browser/lib/auto-instrumentation/index.ts
@@ -1,4 +1,3 @@
 export * from './full-page-load-plugin'
-export * from './network-request-plugin'
 export * from './resource-load-plugin'
 export * from './route-change-plugin'
