+++ packages/platforms/browser/lib/browser.ts
@@ -1,14 +1,14 @@
 import { createClient, InMemoryQueue } from '@bugsnag/core-performance'
-import { FullPageLoadPlugin, NetworkRequestPlugin, ResourceLoadPlugin, RouteChangePlugin } from './auto-instrumentation'
+import createFetchDeliveryFactory from '@bugsnag/delivery-fetch-performance'
+import { createFetchRequestTracker, createXmlHttpRequestTracker, NetworkRequestPlugin } from '@bugsnag/request-tracker-performance'
+import { FullPageLoadPlugin, ResourceLoadPlugin, RouteChangePlugin } from './auto-instrumentation'
 import createBrowserBackgroundingListener from './backgrounding-listener'
 import createClock from './clock'
 import { createSchema } from './config'
 import { createDefaultRoutingProvider } from './default-routing-provider'
-import createFetchDeliveryFactory from '@bugsnag/delivery-fetch-performance'
 import idGenerator from './id-generator'
 import createOnSettle from './on-settle'
 import makeBrowserPersistence from './persistence'
-import { createFetchRequestTracker, createXmlHttpRequestTracker } from '@bugsnag/request-tracker-performance'
 import createResourceAttributesSource from './resource-attributes-source'
 import createSpanAttributesSource from './span-attributes-source'
 import { WebVitals } from './web-vitals'
