+++ packages/request-tracker/lib/index.ts
@@ -2,3 +2,5 @@ export * from './request-tracker'
 export { default as createFetchRequestTracker } from './request-tracker-fetch'
 export { default as createXmlHttpRequestTracker } from './request-tracker-xhr'
 export { default as getAbsoluteUrl } from './url-helpers'
+export * from './auto-instrumentation/network-request-callback'
+export * from './auto-instrumentation/network-request-plugin'
